chore: organize views into separate files
